mixin structure reorganizing
