+++ gradle.properties
@@ -105,7 +105,7 @@ mixin_booter_version = 9.3
 # You should change package root once they are generated
 generate_mixins_json = true
 # Delimit configs with spaces. Should only put configs name instead of full file name
-mixin_configs = thaumicspeedup thaumicspeedup_bwmcompat
+mixin_configs = thaumcraft betterwithmods
 mixin_config_template = thaumicspeedup.mixins.%s.json
 # A refmap is a json that denotes mapping conversions, this json is generated automatically, with the name `mixins.mod_id.refmap.json`
 # Use the property `mixin_refmap` if you want it to use a different name, only one name is accepted
