+++ src/main/resources/thaumicspeedup.mixins.betterwithmods.json
@@ -1,5 +1,5 @@
 {
-	"package": "zone.rong.thaumicspeedup.mixins.bwm",
+	"package": "zone.rong.thaumicspeedup.mixins.betterwithmods",
 	"required": true,
 	"refmap": "mixins.thaumicspeedup.refmap.json",
 	"target": "@env(DEFAULT)",
