+++ tags.properties
@@ -2,3 +2,4 @@ VERSION = ${mod_version}
 MOD_ID = ${mod_id}
 MOD_NAME = ${mod_name}
 MIXIN_CONFIG_TEMPLAE = ${mixin_config_template}
+MIXIN_CONFIGS = ${mixin_configs}
